--- a/lib/orm/associations/has-many.js
+++ b/lib/orm/associations/has-many.js
@@ -81,8 +81,8 @@ export default class HasMany extends Association {
 
     // TODO: look how this is used. Are these necessary, seems like they could be gotten from the above?
     // Or we could use a single data structure to store this information?
-    modelPrototype.associationKeys.push(key);
-    modelPrototype.associationIdKeys.push(foreignKey);
+    modelPrototype.associationKeys.add(key);
+    modelPrototype.associationIdKeys.add(foreignKey);
 
     Object.defineProperty(modelPrototype, foreignKey, {
       /*