--- a/lib/orm/model.js
+++ b/lib/orm/model.js
@@ -161,8 +161,8 @@ class Model {
 
     Object.keys(attrs).forEach(function(attr) {
       if (
-        !this.associationKeys.includes(attr) &&
-        !this.associationIdKeys.includes(attr)
+        !this.associationKeys.has(attr) &&
+        !this.associationIdKeys.has(attr)
       ) {
         this._definePlainAttribute(attr);
       }
@@ -530,8 +530,7 @@ class Model {
 
     Object.keys(attrs).forEach(attr => {
       const isAssociation =
-        this.associationKeys.includes(attr) ||
-        this.associationIdKeys.includes(attr);
+        this.associationKeys.has(attr) || this.associationIdKeys.has(attr);
 
       if (!isAssociation) {
         this.attrs[attr] = attrs[attr];
@@ -586,9 +585,8 @@ class Model {
   _setupRelationships(attrs) {
     Object.keys(attrs).forEach(attr => {
       const value = attrs[attr];
-      const isFk =
-        this.associationIdKeys.includes(attr) || this.fks.includes(attr);
-      const isAssociation = this.associationKeys.includes(attr);
+      const isFk = this.associationIdKeys.has(attr) || this.fks.includes(attr);
+      const isAssociation = this.associationKeys.has(attr);
 
       if (isFk) {
         if (value !== undefined && value !== null) {
@@ -613,7 +611,7 @@ class Model {
 
       // Verify attrs passed in for associations are actually associations
       {
-        if (this.associationKeys.includes(key)) {
+        if (this.associationKeys.has(key)) {
           let association = this.associationFor(key);
           let isNull = value === null;
 
@@ -640,7 +638,7 @@ class Model {
 
       // Verify attrs passed in for association foreign keys are actually fks
       {
-        if (this.associationIdKeys.includes(key)) {
+        if (this.associationIdKeys.has(key)) {
           if (key.endsWith("Ids")) {
             let isArray = Array.isArray(value);
             let isNull = value === null;
@@ -667,7 +665,7 @@ class Model {
           let modelOrCollection = attrs[key];
 
           assert(
-            this.associationKeys.includes(key),
+            this.associationKeys.has(key),
             `You're trying to create a ${
               this.modelName
             } model and you passed in a ${modelOrCollection.toString()} under the ${key} key, but you haven't defined that key as an association on your model.`