--- a/lib/orm/schema.js
+++ b/lib/orm/schema.js
@@ -89,8 +89,8 @@ export default class Schema {
     // Set up associations
     ModelClass.prototype.hasManyAssociations = {}; // a registry of the model's hasMany associations. Key is key from model definition, value is association instance itself
     ModelClass.prototype.belongsToAssociations = {}; // a registry of the model's belongsTo associations. Key is key from model definition, value is association instance itself
-    ModelClass.prototype.associationKeys = []; // ex: address.user, user.addresses
-    ModelClass.prototype.associationIdKeys = []; // ex: address.user_id, user.address_ids
+    ModelClass.prototype.associationKeys = new Set(); // ex: address.user, user.addresses
+    ModelClass.prototype.associationIdKeys = new Set(); // ex: address.user_id, user.address_ids
     ModelClass.prototype.dependentAssociations = []; // a registry of associations that depend on this model, needed for deletion cleanup.
 
     let fksAddedFromThisModel = {};