--- a/lib/serializers/json-api-serializer.js
+++ b/lib/serializers/json-api-serializer.js
@@ -307,7 +307,9 @@ class JSONAPISerializer extends Serializer {
   }
 
   _maybeAddRelationshipsToResourceObjectForModel(hash, model) {
-    let relationships = model.associationKeys.reduce((relationships, key) => {
+    const relationships = {};
+
+    model.associationKeys.forEach(key => {
       let relationship = model[key];
       let relationshipKey = this.keyForRelationship(key);
       let relationshipHash = {};
@@ -343,9 +345,7 @@ class JSONAPISerializer extends Serializer {
       if (!isEmpty(relationshipHash)) {
         relationships[relationshipKey] = relationshipHash;
       }
-
-      return relationships;
-    }, {});
+    });
 
     if (!isEmpty(relationships)) {
       hash.relationships = relationships;
@@ -448,7 +448,7 @@ class JSONAPISerializer extends Serializer {
         let relationshipKey = relationshipKeys[0];
         let graphRelationshipKey = relationshipKey;
         let normalizedRelationshipKey = camelize(relationshipKey);
-        let hasAssociation = model.associationKeys.includes(
+        let hasAssociation = model.associationKeys.has(
           normalizedRelationshipKey
         );