--- a/lib/server.js
+++ b/lib/server.js
@@ -725,8 +725,9 @@ export default class Server {
 
     let list = [];
 
+    const buildArgs = [type, ...traitsAndOverrides];
     for (let i = 0; i < amount; i++) {
-      list.push(this.build(type, ...traitsAndOverrides));
+      list.push(this.build.apply(this, buildArgs));
     }
 
     return list;
@@ -886,8 +887,10 @@ export default class Server {
       : `_${this.inflector.pluralize(type)}`;
     let collection = this.db[collectionName];
 
+    const createArguments = [type, ...traitsAndOverrides, collection];
+
     for (let i = 0; i < amount; i++) {
-      list.push(this.create(type, ...traitsAndOverrides, collection));
+      list.push(this.create.apply(this, createArguments));
     }
 
     return list;