@samselikoff 2019-10-04
Awesome, Set seems good. I'd love to have automated ie11 coverage bc I wouldn't want to regress but it does seem to be supported.